Reduce repeated code for same observers
